+++ packages/stream_chat/test/src/core/models/channel_test.dart
@@ -61,7 +61,6 @@ void main() {
 
     expect(channel.hidden, false);
     expect(channel.extraData['hidden'], false);
-    print(channel.toJson());
     expect(channel.toJson(), {
       'id': 'cid',
       'type': 'test',
@@ -105,7 +104,6 @@ void main() {
 
     expect(channel.disabled, false);
     expect(channel.extraData['disabled'], false);
-    print(channel.toJson());
     expect(channel.toJson(), {
       'id': 'cid',
       'type': 'test',
@@ -150,7 +148,6 @@ void main() {
 
     expect(channel.truncatedAt, currentDate);
     expect(channel.extraData['truncated_at'], currentDate.toIso8601String());
-    print(channel.toJson());
     expect(channel.toJson(), {
       'id': 'cid',
       'type': 'test',