+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+- Added support for `Message.restrictedVisibility` field.
+
 ## 9.4.0
 
 - Updated minimum Flutter version to 3.27.4 for the SDK.